# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Arcology API
  version: "1.0"
  description: |
    REST API for Arcology, a digital artefact catalogue for retrocomputing collections.

    ## Authentication

    All endpoints except `/health` require an API key. Pass it in one of two ways:

    - `Authorization: Bearer <key>` header
    - `X-API-Key: <key>` header

    User API keys are created via the web UI (user profile page). The analysis
    worker uses a separate pre-shared secret configured as `WORKER_API_KEY`.

    ### Permission levels

    | Level | Allowed operations |
    |---|---|
    | `read_only` | GET requests |
    | `read_upload` | GET + create items, artefacts, and analysis jobs |
    | `read_write` | Full access (GET + POST + PUT + PATCH + DELETE) |

    Endpoints marked **Worker only** additionally require the pre-shared
    `WORKER_API_KEY` rather than a user API key.
servers:
  - url: /api
    description: Arcology REST API
security:
  - ApiKeyHeader: []
  - BearerToken: []
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed in the `X-API-Key` header.
    BearerToken:
      type: http
      scheme: bearer
      description: API key passed as a Bearer token in the `Authorization` header.
  responses:
...
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: API key lacks the required permission level.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    FilesystemType:
      type: string
      enum:
        - fat12
        - fat16
        - fat32
        - ntfs
        - hpfs
        - hfs
        - hfs_plus
        - adfs
        - dfs
        - amiga_ofs
        - amiga_ffs
        - iso9660
        - cdfs
        - cpm
        - archive
        - unknown
        - other
...
    ArtefactWithPartitions:
      allOf:
        - $ref: '#/components/schemas/Artefact'
        - type: object
          properties:
            partitions:
              type: array
              items:
                $ref: '#/components/schemas/Partition'

    # --- Analysis ---
...
    Partition:
      type: object
      properties:
        id:
          type: integer
        uuid:
          type: string
        partition_index:
          type: integer
          description: Zero-based index of this partition within the artefact.
        label:
          type: string
          nullable: true
        filesystem:
          $ref: '#/components/schemas/FilesystemType'
        container_format:
          type: string
          nullable: true
          description: Human-readable format description, e.g. `Acorn ADFS E`.
        total_files:
          type: integer
          nullable: true
        unique_files:
          type: integer
          nullable: true
...
paths:
...
  /artefacts/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: get_artefact
      summary: Get artefact
      description: Returns the artefact including its list of partitions.
      tags: [Artefacts]
      responses:
        "200":
          description: Artefact with partitions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtefactWithPartitions'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: update_artefact
      summary: Update artefact checksums
      description: Update the `md5` and/or `sha256` fields on an artefact.
      tags: [Artefacts]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                md5:
                  type: string
                sha256:
                  type: string
      responses:
        "200":
          description: Updated artefact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artefact'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: delete_artefact
      summary: Delete artefact
      description: Permanently deletes the artefact and removes its file from storage.
      tags: [Artefacts]
      responses:
        "204":
          description: Deleted successfully.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /artefacts/{uuid}/partitions:
    parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
        description: Artefact UUID.
    post:
      operationId: add_partition
      summary: Add partition to artefact
      description: >
        Register a filesystem partition found within an artefact.
        **Worker only** — requires the pre-shared `WORKER_API_KEY`.

        Idempotent: any existing partition with the same `partition_index` is
        replaced before the new one is inserted.
      tags: [Partitions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                partition_index:
                  type: integer
                  default: 0
                label:
                  type: string
                filesystem:
                  $ref: '#/components/schemas/FilesystemType'
                container_format:
                  type: string
                  description: Human-readable format, e.g. `Acorn ADFS E`.
                total_files:
                  type: integer
                total_bytes:
                  type: integer
              required: [filesystem]
      responses:
        "201":
          description: Partition created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partition'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /artefacts/{uuid}/analysis:
    parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: list_artefact_analyses
      summary: List analyses for an artefact
      tags: [Analysis]
      responses:
        "200":
          description: List of analyses.
          content:
            application/json:
              schema:
                type: object
                properties:
                  analyses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Analysis'
                required: [analyses]
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      operationId: request_analysis
      summary: Request an analysis job
      description: >
        Queue a new analysis job for an artefact. If an identical PENDING or
        RUNNING job already exists (same type and hints), the existing job is
        returned instead of creating a duplicate (idempotent).
      tags: [Analysis]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                analysis_type:
                  $ref: '#/components/schemas/AnalysisType'
                tool_name:
                  type: string
                hints:
                  type: string
                  description: >
                    JSON-encoded hints for the worker,
                    e.g. `{"platform": "bbc_micro"}`.
              required: [analysis_type]
      responses:
        "200":
          description: Existing job returned (idempotent — job already pending or running).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analysis'
        "201":
          description: New analysis job created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analysis'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /analysis/pending:
    get:
      operationId: get_pending_analyses
      summary: List pending analyses
      description: >
        Returns up to 50 analyses with status `pending`, ordered by creation
        time. Used by workers to poll for new jobs. Each result includes compact
        artefact context so the worker can locate the file.
      tags: [Analysis]
      responses:
        "200":
          description: Pending analyses with artefact context.
          content:
            application/json:
              schema:
                type: object
                properties:
                  analyses:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalysisWithArtefact'
                required: [analyses]
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /analysis/reset-stale:
    post:
      operationId: reset_stale_analyses
      summary: Reset stale running analyses
      description: >
        Resets any analyses that have been in `running` state longer than the
        configured stale-job timeout (default 1 hour) back to `pending`, so
        they can be picked up again. Used by workers on startup to recover from
        crashes, and by operators via the UI.
      tags: [Analysis]
      responses:
        "200":
          description: Number of jobs reset.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reset:
                    type: integer
                    description: Number of stale analyses reset to pending.
                required: [reset]
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /analysis/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
        description: Analysis UUID (32-character hex string).
    get:
      operationId: get_analysis
      summary: Get analysis by UUID
      tags: [Analysis]
      responses:
        "200":
          description: Analysis object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analysis'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /analysis/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Analysis integer ID.
    put:
      operationId: update_analysis
      summary: Update analysis (worker)
      description: >
        Update the status and results of an analysis job. Most result fields
        (`output_path`, `details`, `success`, `summary`, `error_message`,
        `tool_name`, `tool_version`, `output_url`) require the pre-shared
        `WORKER_API_KEY`; ordinary `read_write` keys may only update `status`.

        To atomically claim a pending job, set `claim_worker: true` with
        `status: running`. Only one worker will succeed; others receive 409.
      tags: [Analysis]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/AnalysisStatus'
                tool_name:
                  type: string
                tool_version:
                  type: string
                output_url:
                  type: string
                output_path:
                  type: string
                  description: "**Worker only.** Relative path to output file."
                success:
                  type: boolean
                  description: "**Worker only.**"
                summary:
                  type: string
                  description: "**Worker only.** Short human-readable result summary."
                details:
                  type: string
                  description: "**Worker only.** JSON-encoded structured results."
                error_message:
                  type: string
                  description: "**Worker only.**"
                claim_worker:
                  type: boolean
                  description: >
                    If true, atomically claims the job (transitions pending →
                    running). Returns 409 if already claimed by another worker.
      responses:
        "200":
          description: Analysis updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Analysis'
                  - type: object
                    properties:
                      claimed:
                        type: boolean
                        description: >
                          Present when `claim_worker` was sent. True if this
                          worker successfully claimed the job.
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Job already claimed by another worker.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /hash-databases/recognition-config:
    get:
      operationId: hash_database_recognition_config
      summary: Get recognition configuration
      description: >
        Returns full product and file-hash data for all databases that have
        `enable_product_recognition` set to true. Used by workers to load
        matching rules at startup.
      tags: [Hash Databases]
      responses:
        "200":
          description: Recognition configuration for all enabled databases.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecognitionConfig'
        "401":
          $ref: '#/components/responses/Unauthorized'
...